+++ pdf/lib/widgets/text.dart
@@ -20,6 +20,8 @@ part of widget;
 
 enum TextAlign { left, right, center, justify }
 
+enum TextDirection { ltr, rtl }
+
 abstract class _Span {
   _Span(this.style);
 
@@ -443,12 +445,14 @@ class RichText extends Widget {
   RichText(
       {@required this.text,
       TextAlign textAlign,
+      TextDirection textDirection,
       bool softWrap,
       this.tightBounds = false,
       this.textScaleFactor = 1.0,
       int maxLines})
       : assert(text != null),
         _textAlign = textAlign,
+        textDirection = textDirection ?? TextDirection.ltr,
         _softWrap = softWrap,
         _maxLines = maxLines;
 
@@ -458,6 +462,8 @@ class RichText extends Widget {
 
   TextAlign get textAlign => _textAlign;
   TextAlign _textAlign;
+
+  final TextDirection textDirection;
 
   final double textScaleFactor;
 
@@ -503,6 +509,17 @@ class RichText extends Widget {
           x += delta;
         }
         return totalWidth;
+    }
+
+    if (textDirection == TextDirection.rtl) {
+      for (_Span span in spans) {
+        span.offset = PdfPoint(
+          totalWidth - (span.offset.x + span.width) - delta,
+          span.offset.y - baseline,
+        );
+      }
+
+      return totalWidth;
     }
 
     for (_Span span in spans) {
@@ -566,7 +583,11 @@ class RichText extends Widget {
         final PdfFontMetrics space =
             font.stringMetrics(' ') * (style.fontSize * textScaleFactor);
 
-        final List<String> spanLines = span.text.split('\n');
+        final List<String> spanLines = (textDirection == TextDirection.rtl
+                ? PdfArabic.convert(span.text)
+                : span.text)
+            .split('\n');
+
         for (int line = 0; line < spanLines.length; line++) {
           for (String word in spanLines[line].split(RegExp(r'\s'))) {
             if (word.isEmpty) {
@@ -846,6 +867,7 @@ class Text extends RichText {
     String text, {
     TextStyle style,
     TextAlign textAlign,
+    TextDirection textDirection,
     bool softWrap,
     bool tightBounds = false,
     double textScaleFactor = 1.0,
@@ -856,6 +878,7 @@ class Text extends RichText {
             textAlign: textAlign,
             softWrap: softWrap,
             tightBounds: tightBounds,
+            textDirection: textDirection,
             textScaleFactor: textScaleFactor,
             maxLines: maxLines);
 }